requirements.txt
Flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
pyinstaller==6.3.0
gspread==6.0.0
google-auth==2.27.0
